+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 ## Master Dev Branch
+- Delete all entries from map
+- Don't add map entry if already exists
+- Handle multi line entries in add custom map entry
+- Add missing updated_at from custom map search
 
 ## v1.6.5 - Released: 2025-10-01